html2canvas システム完成
基本的なjs読み込み、機能の完成
logic.jsファイルを作成し、html2canvasファイルの設定を記述
