+++ src/views/App/MemoPrint/MemoPrint.js
@@ -3,37 +3,10 @@ import ReactDOM from 'react-dom';
 import classNames from 'classnames';
 import { Row, Col } from 'reactstrap';
 import { rgbToHex } from '@coreui/coreui/dist/js/coreui-utilities';
-import html2canvas from "html2canvas";
+import { getDisplayImage } from "./logic";
 
 
 class MemoPrint extends Component {
-  getDisplayImage() {
-    //html2canvas実行
-    html2canvas(document.getElementById("target")).then(function(canvas) {
-        this.downloadImage(canvas.toDataURL());
-    });
-  }
-    
-  downloadImage (data) {
-    const fname ="download.png";
-    const encdata= atob(data.replace(/^.*,/, ''));
-    const outdata = new Uint8Array(encdata.length);
-  
-    for (const i = 0; i < encdata.length; i++) {
-        outdata[i] = encdata.charCodeAt(i);
-    }
-    
-    const blob = new Blob([outdata], ["image/png"]);
-    if (window.navigator.msSaveBlob) {
-        //IE用
-        window.navigator.msSaveOrOpenBlob(blob, fname);
-    } else {
-        //それ以外？
-        document.getElementById("getImage").href=data; //base64そのまま設定
-        document.getElementById("getImage").download=fname; //ダウンロードファイル名設定
-        document.getElementById("getImage").click(); //自動クリック
-    }
-  }
 
   render() {
     return (
@@ -43,15 +16,19 @@ class MemoPrint extends Component {
             <i className="icon-drop"></i> Memo print
           </div>
           <div className="card-body" id="target">
-            <p>aaaaaaaaaaaa</p>
+            <p>aadfv</p>
 
-            <div id="downloadImageButton" class="btn btn-primary btn-lg" onClick={this.getDisplayImage()} >Sava image</div>
-            <a id="getImage" href=""  download="image.png">画像保存</a>
+            <div id="downloadImageButton" className="btn btn-primary btn-lg" onClick={this.handleClick} >Sava image</div>
+            <a id="getImage" href="" className="d-none" download="image.png">画像保存</a>
           </div>
         </div>
       </div>
     );
   }
+
+  handleClick = () => {
+    getDisplayImage();
+  };
 }
 
 export default MemoPrint;
